requirements.txt
streamlit==1.24.0
tensorflow==2.10.0
numpy==1.23.5
pillow==9.5.0
nltk==3.8.1
protobuf==3.19.6